requirements.txt
# Core blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==4.0.0
eth-abi==4.2.1
hexbytes==0.3.1

# API clients
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Dashboard (Streamlit + graphiques)
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
cachetools==5.3.2
colorlog==6.8.0
psutil==5.9.6